added api endpoint tests AAA
