+++ package.json
@@ -11,6 +11,10 @@
   "license": "ISC",
   "dependencies": {
     "body-parser": "^1.20.2",
-    "express": "^4.18.2"
+    "express": "^4.18.2",
+    "jest": "^29.7.0"
+  },
+  "devDependencies": {
+    "supertest": "^6.3.4"
   }
 }
